Support Hadoop-3.2
