+++ client-mr/core/src/main/java/org/apache/hadoop/mapred/RssMapOutputCollector.java
