+++ client-mr/core/src/main/java/org/apache/hadoop/mapreduce/RssMRUtils.java
