+++ client-mr/core/src/main/java/org/apache/hadoop/mapreduce/task/reduce/RssFetcher.java
